requirements.txt
faker
fastapi
uvicorn
fastapi_pagination
pytest
apache-airflow
apache-airflow-providers-postgres
pandas
psycopg2-binary